requirements.txt
# install with 'pip install -r requirements.txt'

Flask
Flask-Script
Flask-SQLAlchemy
Flask-Login
gunicorn
Flask-WTF
psycopg2
